Fix spec race condition

+++ backend/spec/helpers/motd.helper.js
@@ -2,37 +2,49 @@ const mongo = require('mongodb');
 const MongoClient = require('mongodb').MongoClient;
 
 
-module.exports = class DatabaseHelper {
+module.exports = class MotdHelper {
   constructor() { }
 
   async connect() {
-    this.database = new MongoClient(
-      'mongodb://mongo:27017', {
-        useNewUrlParser: true,
-        useUnifiedTopology: true
-      }
-    )
-    return this.database.connect((err, client) => {
-      if (err) throw err;
-      this.database = client.db("app");
-      this.motds = this.database.collection("motds")
+    this.client = new MongoClient('mongodb://mongo:27017', {
+      useNewUrlParser: true,
+      useUnifiedTopology: true
     });
-  }
-
-  async clean() {
-    return this.motd.deleteMany({ })
+    try {
+      await this.client.connect();
+      this.motds = this.client.db("app").collection("motds");
+      return Promise.resolve(this.motds);
+    } catch (e) {
+        console.error(e);
+    }
   }
 
   async create(motd) {
-    return this.motds.insertOne({
-      message : motd.message,
-      foreground : motd.foreground,
-      background : motd.background,
-      timestamp : motd.timestamp
-    });
+    if (this.client.isConnected) {
+      return this.motds.insertOne({
+        message : motd.message,
+        foreground : motd.foreground,
+        background : motd.background,
+        timestamp : motd.timestamp
+      });
+    } else {
+      throw PromiseRejectionEvent
+    }
   }
 
   async remove(id) {
-    return this.motds.deleteOne({ _id: mongo.ObjectID(id) });
+    if (this.client.isConnected) {
+      return this.motds.deleteOne({ _id: mongo.ObjectID(id) });
+    } else {
+        throw PromiseRejectionEvent
+    }
+  }
+
+  async removeAll() {
+    if (this.client.isConnected) {
+      return this.motds.deleteMany({ })
+    } else {
+      throw PromiseRejectionEvent
+    }
   }
 }
